requirements.txt
asgiref==3.7.2
Django==4.0.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.13.1
djongo==1.3.6
mongoengine==0.23.1
pymongo==3.12.3
pytz==2021.3
sqlparse==0.2.4
typing_extensions==4.9.0
tzdata==2021.5
